wikipedia/requirements_with_tf.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
isort==4.3.15
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
pkg-resources==0.0.0
protobuf==3.7.0
py4j==0.10.7
pylint==2.3.1
pyspark==2.4.1
pyspark-stubs==2.4.0.post1
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
torch==1.0.1.post2
typed-ast==1.3.1
url-normalize==1.4.1
Werkzeug==0.15.2
wrapt==1.11.1
